+++ lib/t/t_ext2int2ext.m
@@ -15,7 +15,7 @@ if nargin < 1
     quiet = 0;
 end
 
-num_tests = 138;
+num_tests = 165;
 t_begin(num_tests, quiet);
 
 %% define named indices into data matrices
@@ -194,6 +194,42 @@ else
     t_is(mpc.N, mpci.N, 12, [t 'N']);
     t = 'mpc = int2ext(mpc) : ';
     mpc = int2ext(mpc);
+    t_is(mpc.bus, mpce.bus, 12, [t 'bus']);
+    t_is(mpc.branch, mpce.branch, 12, [t 'branch']);
+    t_is(mpc.gen, mpce.gen, 12, [t 'gen']);
+    t_is(mpc.gencost, mpce.gencost, 12, [t 'gencost']);
+    t_ok(isequaln(mpc.bus_name, mpce.bus_name), [t 'bus_name']);
+    t_ok(isequaln(mpc.gentype, mpce.gentype), [t 'gentype']);
+    t_ok(isequaln(mpc.genfuel, mpce.genfuel), [t 'genfuel']);
+    t_is(mpc.A, mpce.A, 12, [t 'A']);
+    t_is(mpc.N, mpce.N, 12, [t 'N']);
+
+    %%-----  mpc = ext2int/int2ext(mpc, mpopt)  -----
+    mpopt = mpoption();
+    t = 'mpc = ext2int(mpc, mpopt) : ';
+    mpc = ext2int(mpce, mpopt);
+    t_is(mpc.bus, mpci.bus, 12, [t 'bus']);
+    t_is(mpc.branch, mpci.branch, 12, [t 'branch']);
+    t_is(mpc.gen, mpci.gen, 12, [t 'gen']);
+    t_is(mpc.gencost, mpci.gencost, 12, [t 'gencost']);
+    t_ok(isequaln(mpc.bus_name, mpci.bus_name), [t 'bus_name']);
+    t_ok(isequaln(mpc.gentype, mpci.gentype), [t 'gentype']);
+    t_ok(isequaln(mpc.genfuel, mpci.genfuel), [t 'genfuel']);
+    t_is(mpc.A, mpci.A, 12, [t 'A']);
+    t_is(mpc.N, mpci.N, 12, [t 'N']);
+    t = 'mpc = ext2int(mpc, mpopt) - repeat : ';
+    mpc = ext2int(mpc, mpopt);
+    t_is(mpc.bus, mpci.bus, 12, [t 'bus']);
+    t_is(mpc.branch, mpci.branch, 12, [t 'branch']);
+    t_is(mpc.gen, mpci.gen, 12, [t 'gen']);
+    t_is(mpc.gencost, mpci.gencost, 12, [t 'gencost']);
+    t_ok(isequaln(mpc.bus_name, mpci.bus_name), [t 'bus_name']);
+    t_ok(isequaln(mpc.gentype, mpci.gentype), [t 'gentype']);
+    t_ok(isequaln(mpc.genfuel, mpci.genfuel), [t 'genfuel']);
+    t_is(mpc.A, mpci.A, 12, [t 'A']);
+    t_is(mpc.N, mpci.N, 12, [t 'N']);
+    t = 'mpc = int2ext(mpc, mpopt) : ';
+    mpc = int2ext(mpc, mpopt);
     t_is(mpc.bus, mpce.bus, 12, [t 'bus']);
     t_is(mpc.branch, mpce.branch, 12, [t 'branch']);
     t_is(mpc.gen, mpce.gen, 12, [t 'gen']);